refactor: put operators in the same file
