+++ providers/docling/provider.yaml
@@ -41,5 +41,4 @@ connection-types:
 operators:
   - integration-name: Docling
     python-modules:
-      - airflow.providers.docling.operators.docling_process
-      - airflow.providers.docling.operators.docling_upload_source
+      - airflow.providers.docling.operators.docling_convert
